Fix fatal SSH starting error & correct comment
Fixed a fatal starting error that occurred when obtaining the example server-private-key value. this was removed for another reason, instead of being fixed:
using the system SSH rsa key could be considered bad practice due to the insecurities of using the same key twice. in the event someone (maybe by the web gui) gets the key, that same key can be used as a system console assuming SSH was enabled on the host machine.