Bump mc-server-lib to version 3.7.0 and update 'software' key comments in UpdaterConfig.java
This commit bumps the mc-server-lib library to version 3.7.0 which adds support for updating PandaSpigot and Mohist. There were some removed server software and some added ones in the previous release. The comments of the 'software' key in the UpdaterConfig.java have also been updated accordingly to the up-to-date supported updaters and are now in the order as it is in the UpdateBuilder.java file of the MCServerUpdater project. A little space was also missing in the comments (which caused a space to miss after a comma), this has also been fixed accordingly.

Signed-off-by: Minionguyjpro <minionguyjpro@gmail.com>